Got all data points working on the MPU

+++ trunk/timbot/experiments/Sensor testing/IR_analog/IR_analog.ino
@@ -6,7 +6,7 @@
 #define SAMPLE_AMOUNT 500
 
 int sensorPin = A0;
-int sensorValue = 0;
+long sensorValue = 0;
 
 void setup() {
   Serial.begin(9600);
